requirements.txt
pyspellchecker
pandas